refactor(scan) move test utils to test_utils.rs
