+++ crates/iceberg/src/inspect/manifests.rs
@@ -289,7 +289,7 @@ mod tests {
     use expect_test::expect;
     use futures::TryStreamExt;
 
-    use crate::scan::tests::TableTestFixture;
+    use crate::scan::test_utils::TableTestFixture;
     use crate::spec::TableMetadata;
     use crate::test_utils::check_record_batches;
